refactor(innertube): modularize innertube utilities
